video-tools/requirements.txt
python-dotenv
requests==2.32.5
jsonschema==4.25.1
httpx
pydantic-ai==1.0.10
claude-agent-sdk
Pillow
boto3
motor
